refactor: move edit transactions to financial module
