+++ apps/dashboard/src/modules/financial/components/transaction/TransactionAmountsConfirmDialog.vue
