+++ apps/dashboard/src/modules/financial/components/transaction/TransactionDetailCard.vue
