+++ apps/dashboard/src/modules/financial/components/transaction/TransactionEditCard.vue
@@ -39,8 +39,8 @@ import TransactionEditConfirmDialog from './TransactionEditConfirmDialog.vue';
 import FormCard from '@/components/FormCard.vue';
 import { updateTransactionUserObject } from '@/utils/validation-schema';
 import { schemaToForm, getProperty } from '@/utils/formUtils';
-import { useTransactionCard } from '@/composables/transactionCard';
-import { useTransactionForm } from '@/composables/transactionForm';
+import { useTransactionCard } from '../../composables/useTransactionCard';
+import { useTransactionForm } from '../../composables/useTransactionForm';
 
 const { t } = useI18n();
 const props = defineProps<{ transactionId: number }>();
