+++ apps/dashboard/src/modules/financial/components/transaction/forms/TransactionAmountsBaseForm.vue
