+++ apps/dashboard/src/modules/financial/views/transaction/TransactionView.vue
@@ -13,8 +13,8 @@ import { useI18n } from 'vue-i18n';
 import { useRoute } from 'vue-router';
 import type { AxiosError } from 'axios';
 import PageContainer from '@/layout/PageContainer.vue';
-import TransactionSearchCard from '@/modules/admin/components/transactions/TransactionSearchCard.vue';
-import TransactionDetailCard from '@/modules/admin/components/transactions/TransactionDetailCard.vue';
+import TransactionSearchCard from '@/modules/financial/components/transaction/TransactionSearchCard.vue';
+import TransactionDetailCard from '@/modules/financial/components/transaction/TransactionDetailCard.vue';
 import { useTransactionStore } from '@/stores/transaction.store';
 import apiService from '@/services/ApiService';
 
@@ -44,30 +44,25 @@ const currentTransaction = computed(() => {
 });
 
 async function handleSearch(id: number) {
-  errorMessage.value = '';
-  currentTransactionId.value = null;
-
   try {
+    errorMessage.value = '';
+    currentTransactionId.value = id;
     await transactionStore.fetchIndividualTransaction(id, apiService);
-    currentTransactionId.value = id; // Set the ID to trigger reactive computation
   } catch (error) {
     const axiosError = error as AxiosError;
     if (axiosError.response?.status === 404) {
       errorMessage.value = t('modules.admin.transactions.notFound');
     } else {
-      errorMessage.value = t('common.apiError');
+      errorMessage.value = t('common.error.generic');
     }
+    currentTransactionId.value = null;
   }
 }
 
-// Check for transaction ID in route query and auto-search
 onMounted(() => {
-  const transactionId = route.query.id;
-  if (transactionId && typeof transactionId === 'string') {
-    const id = parseInt(transactionId, 10);
-    if (!isNaN(id)) {
-      void handleSearch(id);
-    }
+  // If there's an initial transaction ID from the route, search for it
+  if (initialTransactionId.value) {
+    handleSearch(initialTransactionId.value);
   }
 });
 </script>
